+++ MicroBlink.framework/Headers/PPAusIDCombinedRecognizerSettings.h
@@ -1,8 +1,8 @@
 //
-//  PPGermanIDCombinedRecognizerSettings.h
-//  PhotoPayDev
+//  PPAusIDCombinedRecognizerSettings.h
+//  MicroBlinkDev
 //
-//  Created by Jura Skrlec on 22/05/2017.
+//  Created by Jura Skrlec on 19/06/2017.
 //
 //
 
@@ -16,7 +16,8 @@ NS_ASSUME_NONNULL_BEGIN
  * German Combined ID recognizer is used for scanning both front and back side of german IDs.
  */
 PP_CLASS_AVAILABLE_IOS(6.0)
-@interface PPGermanIDCombinedRecognizerSettings : PPCombinedRecognizerSettings
+
+@interface PPAusIDCombinedRecognizerSettings : PPCombinedRecognizerSettings
 
 /**
  * Name of the image sent to didOutputMetadata method of scanDelegate object that contains face.
@@ -90,8 +91,6 @@ PP_CLASS_AVAILABLE_IOS(6.0)
  * Default: NO
  */
 @property (nonatomic, assign) BOOL encodeFullDocumentPhoto;
-
 @end
 
 NS_ASSUME_NONNULL_END
-