+++ MicroBlink.framework/Headers/PPBarDecoderRecognizerResult.h
@@ -12,7 +12,7 @@
 /**
  * Type of the barcode which BarDecoderRecognizer returnes
  */
-typedef NS_ENUM(NSUInteger, PPBarDecoderBarcodeType) {
+typedef PP_DEPRECATED_IOS(1_0, 5_9_3) NS_ENUM(NSUInteger, PPBarDecoderBarcodeType) {
     /** Code 128 */
     PPBarDecoderBarcodeTypeCode128,
     /** Code 39 */
@@ -24,7 +24,7 @@ typedef NS_ENUM(NSUInteger, PPBarDecoderBarcodeType) {
  *
  * Contains raw Barcode detailed data, barcode type, and methods for getting string representation of results.
  */
-PP_CLASS_AVAILABLE_IOS(6.0)
+PP_CLASS_DEPRECATED_IOS(1_0, 5_9_3, "Use PPBarcodeRecognizerResult instead")
 @interface PPBarDecoderRecognizerResult : PPRecognizerResult
 
 /**
@@ -32,7 +32,7 @@ PP_CLASS_AVAILABLE_IOS(6.0)
  *
  *  @return Type of the barcode
  */
-- (PPBarDecoderBarcodeType)barcodeType;
+- (PPBarDecoderBarcodeType)barcodeType PP_DEPRECATED_IOS(1_0, 5_9_3, "Use PPBarcodeType instead");
 
 /**
  * Flag indicating uncertain scanning data
@@ -112,6 +112,6 @@ PP_CLASS_AVAILABLE_IOS(6.0)
  *
  *  @return String representation of a given PPBarDecoderBarcodeType enum value.
  */
-+ (NSString *_Nonnull)toTypeName:(PPBarDecoderBarcodeType)type;
++ (NSString *_Nonnull)toTypeName:(PPBarDecoderBarcodeType)type PP_DEPRECATED_IOS(1_0, 5_9_3, "Use PPBarcodeType instead");
 
 @end