+++ MicroBlink.framework/Headers/PPSingaporeIDCombinedRecognizerSettings.h
@@ -1,8 +1,8 @@
 //
-//  PPGermanIDCombinedRecognizerSettings.h
-//  PhotoPayDev
+//  PPSingaporeIDCombinedRecognizerSettings.h
+//  MicroBlinkDev
 //
-//  Created by Jura Skrlec on 22/05/2017.
+//  Created by Jura Skrlec on 19/06/2017.
 //
 //
 
@@ -16,7 +16,7 @@ NS_ASSUME_NONNULL_BEGIN
  * German Combined ID recognizer is used for scanning both front and back side of german IDs.
  */
 PP_CLASS_AVAILABLE_IOS(6.0)
-@interface PPGermanIDCombinedRecognizerSettings : PPCombinedRecognizerSettings
+@interface PPSingaporeIDCombinedRecognizerSettings : PPCombinedRecognizerSettings
 
 /**
  * Name of the image sent to didOutputMetadata method of scanDelegate object that contains face.
@@ -24,13 +24,6 @@ PP_CLASS_AVAILABLE_IOS(6.0)
  * is enabled via displayFaceImage property and receiving of dewarpedImage in MetadataSettings is enabled.
  */
 + (NSString *)ID_FACE;
-
-/**
- * Name of the image sent to didOutputMetadata method of scanDelegate object that contains signature.
- * This image will be sent to scan delegate during recognition process if displaying of face image
- * is enabled via displaySignatureImage property and receiving of dewarpedImage in MetadataSettings is enabled.
- */
-+ (NSString *)ID_SIGNATURE;
 
 /**
  * Name of the image sent to didOutputMetadata method of scanDelegate object that contains full document front side.
@@ -61,20 +54,6 @@ PP_CLASS_AVAILABLE_IOS(6.0)
  */
 @property (nonatomic, assign) BOOL encodeFacePhoto;
 
-/**
- * Sets whether signature image from ID card should be sent to didOutputMetadata method of scanDelegate object.
- * If you want to recieve this image, be sure to enable dewarpedImage in MetadataSettings.
- *
- * Default: NO
- */
-@property (nonatomic, assign) BOOL returnSignaturePhoto;
-
-/**
- * Sets whether signature image should be encoded in the result object returned by didOutputResult:
- *
- * Default: NO
- */
-@property (nonatomic, assign) BOOL encodeSignaturePhoto;
 
 /**
  * Sets whether full document image of ID card should be sent to didOutputMetadata method of scanDelegate object.
@@ -94,4 +73,3 @@ PP_CLASS_AVAILABLE_IOS(6.0)
 @end
 
 NS_ASSUME_NONNULL_END
-