+++ MicroBlink.framework/Headers/PPSlovakIDFrontRecognizerSettings.h
@@ -80,6 +80,13 @@ PP_CLASS_AVAILABLE_IOS(6.0)
 @property (nonatomic, assign) BOOL extractIssuedBy;
 
 /**
+ * Defines if issuing document number should be extracted from Slovakian ID
+ *
+ * Default: YES
+ */
+@property (nonatomic, assign) BOOL extractDocumentNumber;
+
+/**
  * Sets whether face photo from ID card should be sent to didOutputMetadata method of scanDelegate object.
  * If you want to recieve this image, be sure to enable dewarpedImage in MetadataSettings.
  *